Refactor: Put components in their own folders; put page-style ones into Routes; put all routes in main.tsx
