+++ src/components/DashboardSidebar/DashboardSidebar.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
-import styles from './Dashboard.module.css';
+import styles from '../../routes/Dashboard/Dashboard.module.css';
 import { BsGrid1X2Fill, BsFillArchiveFill, BsFillGrid3X3GapFill, BsMenuButtonWideFill, BsFillGearFill, BsTable } from 'react-icons/bs';
