+++ src/components/SignInSlider/SignInSlider.tsx
