+++ src/main.tsx
@@ -1,18 +1,35 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import Root from './routes/Root/Root.tsx'; 
-import LoginWrapper from './components/Login/LoginWrapper.tsx';
-import App from './components/Dashboard/DashboardApp.tsx';
+import Root from './routes/Root/Root.tsx';
+import LoginPage from './routes/LoginPage/LoginPage.tsx';
+import Dashboard from './routes/Dashboard/Dashboard.tsx';
+import DataPredictions from './components/DashboardDataPredictions/DashboardDataPredictions.tsx';
+import DashboardLanding from './components/DashboardLanding/DashboardLanding.tsx';
 
 const router = createBrowserRouter([
 	{
 		element: <Root />,
 		errorElement: <Root />,
 		children: [
-            
-			{ path: 'login', element: <LoginWrapper /> },
-			{ path: 'dashboard/*', element: <App /> }, // Use App and manage sub-routes inside it
+			{
+				path: 'login',
+				element: <LoginPage />
+			},
+			{
+				path: 'dashboard',
+				element: <Dashboard />,
+				children: [
+					{
+						path: '',
+						element: <DashboardLanding />,
+					},
+					{
+						path: 'data-predictions',
+						element: <DataPredictions />,
+					}
+				]
+			}
 		],
 	},
 ]);
